+++ concordium-sdk/src/main/java/com/concordium/sdk/crypto/encryptedtransfers/AggregatedDecryptedAmount.java
@@ -1,22 +1,16 @@
 package com.concordium.sdk.crypto.encryptedtransfers;
 
 import com.concordium.sdk.responses.accountinfo.AccountEncryptedAmount;
-import com.concordium.sdk.transactions.EncryptedAmount;
-import com.concordium.sdk.transactions.EncryptedAmountIndex;
 import lombok.Builder;
 import lombok.Data;
+import lombok.extern.jackson.Jacksonized;
 
-@Data
+@Jacksonized
 @Builder
-class IndexedEncryptedAmount {
-    /**
-     * The actual encrypted amount.
-     */
-    private final EncryptedAmount encryptedChunks;
-    /**
-     * Index of the amount on the account.
-     */
-    private final EncryptedAmountIndex index;
+@Data
+public class AggregatedDecryptedAmount {
+    private final String encryptedChunks;
+    private final long index;
 
     public static IndexedEncryptedAmount from(AccountEncryptedAmount encryptedAmount) {
         return new IndexedEncryptedAmount(
@@ -24,4 +18,5 @@ class IndexedEncryptedAmount {
                 encryptedAmount.getStartIndex()
         );
     }
+
 }