+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/Transaction.java
@@ -1,7 +1,6 @@
 package com.concordium.sdk.transactions;
 
 import com.concordium.sdk.exceptions.TransactionCreationException;
-import com.concordium.sdk.types.UInt64;
 
 import java.util.Objects;