+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionFactory.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.transactions;
 
+import com.concordium.sdk.crypto.elgamal.ElgamalPublicKey;
 import com.concordium.sdk.crypto.elgamal.ElgamalSecretKey;
 import com.concordium.sdk.crypto.encryptedtransfers.EncryptedTransfers;
 import com.concordium.sdk.responses.accountinfo.AccountEncryptedAmount;
@@ -74,10 +75,22 @@ public class TransactionFactory {
     }
 
 
+    /**
+     * Creates a new {@link TransferToPublicTransaction.TransferToPublicTransactionBuilder} for
+     * creating a {@link TransferToPublicTransaction}
+     *
+     * @return the builder for a {@link TransferToPublicTransaction}
+     */
     public static TransferToPublicTransaction.TransferToPublicTransactionBuilder newTransferToPublic() {
         return TransferToPublicTransaction.builder();
     }
 
+    /**
+     * Creates a new {@link TransferToPublicTransaction.TransferToPublicTransactionBuilder} for
+     * creating a {@link TransferToPublicTransaction}
+     *
+     * @return the builder for a {@link TransferToPublicTransaction}
+     */
     public static TransferToPublicTransaction.TransferToPublicTransactionBuilder newTransferToPublic(
             CryptographicParameters cryptographicParameters,
             AccountEncryptedAmount accountEncryptedAmount,
@@ -97,6 +110,104 @@ public class TransactionFactory {
                 .index(jniOutput.getIndex())
                 .proof(jniOutput.getProof())
                 .remainingAmount(jniOutput.getRemainingAmount());
+    }
+
+    /**
+     * Creates a new {@link TransferToEncryptedTransaction.TransferToEncryptedTransactionBuilder} for
+     * creating a {@link TransferToEncryptedTransaction}
+     *
+     * @return the builder for a {@link TransferToEncryptedTransaction}
+     */
+    public static TransferToEncryptedTransaction.TransferToEncryptedTransactionBuilder newTransferToEncrypted() {
+        return TransferToEncryptedTransaction.builder();
+    }
+
+
+    /**
+     * Creates a new {@link EncryptedTransferTransaction.EncryptedTransferTransactionBuilder} for
+     * creating a {@link EncryptedTransferTransaction}
+     *
+     * @return the builder for a {@link EncryptedTransferTransaction}
+     */
+
+    public static EncryptedTransferTransaction.EncryptedTransferTransactionBuilder newEncryptedTransfer() {
+        return EncryptedTransferTransaction.builder();
+    }
+
+
+    /**
+     * Creates a new {@link EncryptedTransferTransaction.EncryptedTransferTransactionBuilder} for
+     * creating a {@link EncryptedTransferTransaction}
+     *
+     * @return the builder for a {@link EncryptedTransferTransaction}
+     */
+    public static EncryptedTransferTransaction.EncryptedTransferTransactionBuilder newEncryptedTransfer(
+            CryptographicParameters cryptographicParameters,
+            AccountEncryptedAmount accountEncryptedAmount,
+            ElgamalPublicKey receiverPublicKey,
+            ElgamalSecretKey senderSecretKey,
+            CCDAmount amountToSend) {
+        val jniOutput = EncryptedTransfers.createEncryptedTransferPayload(
+                cryptographicParameters,
+                accountEncryptedAmount,
+                receiverPublicKey,
+                senderSecretKey,
+                amountToSend
+        );
+
+        val amountTransferData = EncryptedAmountTransferData.builder()
+                .remainingAmount(jniOutput.getRemainingAmount())
+                .transferAmount(jniOutput.getTransferAmount())
+                .index(jniOutput.getIndex())
+                .proof(jniOutput.getProof()).build();
+
+        return EncryptedTransferTransaction
+                .builder()
+                .data(amountTransferData);
+
+    }
+
+    /**
+     * Creates a new {@link EncryptedTransferWithMemoTransaction.EncryptedTransferWithMemoTransactionBuilder} for
+     * creating a {@link EncryptedTransferWithMemoTransaction}
+     *
+     * @return the builder for a {@link EncryptedTransferWithMemoTransaction}
+     */
+
+    public static EncryptedTransferWithMemoTransaction.EncryptedTransferWithMemoTransactionBuilder newEncryptedTransferWithMemo() {
+        return EncryptedTransferWithMemoTransaction.builder();
+    }
+
+    /**
+     * Creates a new {@link EncryptedTransferWithMemoTransaction.EncryptedTransferWithMemoTransactionBuilder} for
+     * creating a {@link EncryptedTransferWithMemoTransaction}
+     *
+     * @return the builder for a {@link EncryptedTransferWithMemoTransaction}
+     */
+    public static EncryptedTransferWithMemoTransaction.EncryptedTransferWithMemoTransactionBuilder newEncryptedTransferWithMemo(
+            CryptographicParameters cryptographicParameters,
+            AccountEncryptedAmount accountEncryptedAmount,
+            ElgamalPublicKey receiverPublicKey,
+            ElgamalSecretKey senderSecretKey,
+            CCDAmount amountToSend) {
+        val jniOutput = EncryptedTransfers.createEncryptedTransferPayload(
+                cryptographicParameters,
+                accountEncryptedAmount,
+                receiverPublicKey,
+                senderSecretKey,
+                amountToSend
+        );
+
+        val amountTransferData = EncryptedAmountTransferData.builder()
+                .remainingAmount(jniOutput.getRemainingAmount())
+                .transferAmount(jniOutput.getTransferAmount())
+                .index(jniOutput.getIndex())
+                .proof(jniOutput.getProof()).build();
+
+        return EncryptedTransferWithMemoTransaction
+                .builder()
+                .data(amountTransferData);
+
     }
 
     /**